add userService to reuse logic
